+++ clinvar_export/vueapp/patch.sh
@@ -1,0 +1,3 @@
+# Fix https://github.com/capricorn86/happy-dom/issues/548
+# This is the reason why we pinned to happy-dom 6.0.4
+cp HTMLSelectElement.js ./node_modules/happy-dom/lib/nodes/html-select-element/HTMLSelectElement.js