+++ clinvar_export/vueapp/src/main.js
@@ -1,13 +1,9 @@
-import Vue from 'vue'
+import { createPinia } from 'pinia'
+import { createApp } from 'vue'
 
 import App from './App.vue'
-import router from './router'
-import store from './store'
 
-Vue.config.productionTip = false
-
-new Vue({
-  router,
-  store,
-  render: (h) => h(App),
-}).$mount('#app')
+const pinia = createPinia()
+const app = createApp(App)
+app.use(pinia)
+app.mount('#app')