+++ clinvar_export/vueapp/vite.config.js
@@ -1,0 +1,31 @@
+import { fileURLToPath, URL } from 'url'
+const { resolve } = require('path')
+import vue from '@vitejs/plugin-vue'
+import { defineConfig } from 'vite'
+
+// https://vitejs.dev/config/
+export default defineConfig({
+  build: {
+    // generate manifest.json in outDir
+    manifest: true,
+    // overwrite default .html entry
+    outDir: '../../varfish/static/vueapp',
+    rollupOptions: {
+      input: [resolve(__dirname, './src/main.js')],
+    },
+  },
+  plugins: [vue()],
+  resolve: {
+    alias: {
+      '@': fileURLToPath(new URL('./src', import.meta.url)),
+    },
+  },
+  test: {
+    coverage: {
+      all: true,
+      provider: 'istanbul',
+      reporter: ['text', 'json', 'html'],
+    },
+    environment: 'happy-dom',
+  },
+})