+++ HISTORY.rst
@@ -131,6 +131,7 @@ Full Change List
 - Moving clinvar Vue app (#711).
 - Bugfix that allow clinvar export submission set deletion (#713).
 - Removing dependency on bootstrap-vue package (#716).
+- Migrating store dependency from Vuex v3 to Pinia (#720).
 
 ------
 v1.2.0